code/backend/requirements.txt
# Coroutines
celery==5.4.0

# Database
psycopg==3.2.2
psycopg-binary==3.2.2
redis==5.0.8

# Django
Django==5.1.1
django-celery-results==2.5.1
django-simple-history==3.7.0
django-storages==1.14.4
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7

# Logging
sentry-sdk==2.14.0

# Testing
pytest==8.3.3
pytest-cov==5.0.0
pytest-django==4.9.0
pytest-factoryboy==2.7.0

# Media
boto3==1.35.19

# Utils
arrow==1.3.0
flake8==7.1.1
flower==2.0.1
gunicorn==23.0.0
isort==5.13.2
python-decouple==3.8
requests==2.32.3
siwe==4.2.0
web3==6.20.3

# Websockets
channels==4.1.0
channels-redis==4.2.0
daphne==4.1.2
